+++ frontend/src/utils/index.ts
@@ -26,3 +26,4 @@ export * from "./sheet";
 export * from "./v1";
 export * from "./iam";
 export * from "./auto-route";
+export * from "./naive-ui";
