+++ frontend/src/views/sql-editor/EditorPanel/Panels/ExternalTablesPanel/ExternalTableColumnsTable.vue
@@ -29,8 +29,7 @@ import type {
   SchemaMetadata,
   ExternalTableMetadata,
 } from "@/types/proto/v1/database_service";
-import { getHighlightHTMLByRegExp } from "@/utils";
-import { useAutoHeightDataTable } from "../../common";
+import { getHighlightHTMLByRegExp, useAutoHeightDataTable } from "@/utils";
 import { useEditorPanelContext } from "../../context";
 
 const props = defineProps<{
@@ -42,13 +41,9 @@ const props = defineProps<{
 }>();
 
 const { viewState } = useEditorPanelContext();
-const { containerElRef, tableBodyHeight, layoutReady } =
+const { containerElRef, tableBodyHeight, layoutReady, virtualListRef } =
   useAutoHeightDataTable();
 const dataTableRef = ref<DataTableInst>();
-const vlRef = computed(() => {
-  return (dataTableRef.value as any)?.$refs?.mainTableInstRef?.bodyInstRef
-    ?.virtualListRef;
-});
 const { t } = useI18n();
 
 const filteredColumns = computed(() => {
@@ -129,7 +124,7 @@ const columns = computed(() => {
 });
 
 watch(
-  [() => viewState.value?.detail.column, vlRef],
+  [() => viewState.value?.detail.column, virtualListRef],
   ([column, vl]) => {
     if (column && vl) {
       requestAnimationFrame(() => {
