+++ frontend/src/views/sql-editor/EditorPanel/Panels/ProceduresPanel/ProceduresTable.vue
@@ -29,9 +29,8 @@ import type {
   ProcedureMetadata,
   SchemaMetadata,
 } from "@/types/proto/v1/database_service";
-import { getHighlightHTMLByRegExp } from "@/utils";
+import { getHighlightHTMLByRegExp, useAutoHeightDataTable } from "@/utils";
 import { keyWithPosition } from "@/views/sql-editor/EditorCommon";
-import { useAutoHeightDataTable } from "../../common";
 import { useEditorPanelContext } from "../../context";
 
 type ProcedureWithPosition = {
@@ -80,17 +79,18 @@ const filteredProcedures = computed(() => {
   return proceduresWithPosition.value;
 });
 
-const { dataTableRef, containerElRef, tableBodyHeight, layoutReady } =
-  useAutoHeightDataTable(
-    filteredProcedures,
-    computed(() => ({
-      maxHeight: props.maxHeight ? props.maxHeight : null,
-    }))
-  );
-const vlRef = computed(() => {
-  return (dataTableRef.value as any)?.$refs?.mainTableInstRef?.bodyInstRef
-    ?.virtualListRef;
-});
+const {
+  dataTableRef,
+  containerElRef,
+  virtualListRef,
+  tableBodyHeight,
+  layoutReady,
+} = useAutoHeightDataTable(
+  filteredProcedures,
+  computed(() => ({
+    maxHeight: props.maxHeight ? props.maxHeight : null,
+  }))
+);
 
 const columns = computed(() => {
   const columns: (DataTableColumn<ProcedureWithPosition> & {
@@ -128,7 +128,7 @@ const rowProps = ({ procedure, position }: ProcedureWithPosition) => {
 };
 
 watch(
-  [() => viewState.value?.detail.procedure, vlRef],
+  [() => viewState.value?.detail.procedure, virtualListRef],
   ([procedure, vl]) => {
     if (procedure && vl) {
       vl.scrollTo({ key: procedure });
