+++ frontend/src/views/sql-editor/EditorPanel/common.ts
@@ -1,13 +1,12 @@
-import { pausableWatch, useElementSize } from "@vueuse/core";
 import { first } from "lodash-es";
-import type { DataTableInst, SelectOption } from "naive-ui";
-import { computed, ref, unref, watch, type MaybeRef } from "vue";
+import type { SelectOption } from "naive-ui";
+import { computed, watch } from "vue";
 import {
   useConnectionOfCurrentSQLEditorTab,
   useDBSchemaV1Store,
 } from "@/store";
 import { DatabaseMetadataView } from "@/types/proto/v1/database_service";
-import { hasSchemaProperty, nextAnimationFrame } from "@/utils";
+import { hasSchemaProperty } from "@/utils";
 import { useEditorPanelContext } from "./context";
 
 export const useSelectSchema = () => {
@@ -40,66 +39,4 @@ export const useSelectSchema = () => {
   );
 
   return { selectedSchemaName, showSchemaSelect, options };
-};
-
-export type AutoHeightDataTableOptions = {
-  maxHeight: number | null;
-};
-const defaultAutoHeightDataTableOptions = (): AutoHeightDataTableOptions => ({
-  maxHeight: null,
-});
-
-export const useAutoHeightDataTable = (
-  data?: MaybeRef<unknown[]>,
-  options?: MaybeRef<Partial<AutoHeightDataTableOptions>>
-) => {
-  const opts = computed(() => ({
-    ...defaultAutoHeightDataTableOptions(),
-    ...unref(options),
-  }));
-  const GAP_HEIGHT = 2;
-  const dataTableRef = ref<DataTableInst>();
-  const containerElRef = ref<HTMLElement>();
-  const tableHeaderElRef = computed(
-    () =>
-      containerElRef.value?.querySelector("thead.n-data-table-thead") as
-        | HTMLElement
-        | undefined
-  );
-  const { height: containerHeight } = useElementSize(containerElRef);
-  const { height: tableHeaderHeight } = useElementSize(tableHeaderElRef);
-  const tableBodyHeight = ref(0);
-
-  const { pause: pauseWatch, resume: resumeWatch } = pausableWatch(
-    [() => unref(data), containerHeight, tableHeaderHeight, opts],
-    async () => {
-      if (containerHeight.value === 0 || tableHeaderHeight.value === 0) return;
-      if (opts.value.maxHeight === null) {
-        // The table height is always limited by the container height
-        // and need not to be adjusted automatically
-        tableBodyHeight.value =
-          containerHeight.value - tableHeaderHeight.value - GAP_HEIGHT;
-        return;
-      }
-
-      pauseWatch();
-      const experimentalMaxBodyHeight =
-        opts.value.maxHeight - tableHeaderHeight.value - GAP_HEIGHT;
-      tableBodyHeight.value = experimentalMaxBodyHeight;
-      await nextAnimationFrame();
-      resumeWatch();
-    },
-    { immediate: true }
-  );
-
-  // Use this to avoid unnecessary initial rendering
-  const layoutReady = computed(() => tableHeaderHeight.value > 0);
-
-  return {
-    dataTableRef,
-    containerElRef,
-    tableHeaderElRef,
-    tableBodyHeight,
-    layoutReady,
-  };
 };
